Random value generators moved out from developer module
